% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sano_title.R
\name{add_sano_title}
\alias{add_sano_title}
\title{Add a Styled Title to a ggplot}
\usage{
add_sano_title(label, number, subtitle = waiver())
}
\arguments{
\item{label}{A character vector representing the main title text.}

\item{number}{A character or numeric value indicating the figure number. This will be converted to a character.}

\item{subtitle}{A subtitle for the plot. Defaults to \code{waiver()}, which means no subtitle is added.}
}
\value{
A list modifying the plot's title and subtitle, to be used within a ggplot chain.
}
\description{
This function adds a styled title to a ggplot object, incorporating a figure number and custom formatting.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  add_sano_title(label = "Car Weight vs. MPG", number = 1)
print(p)
}
}
