% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sano_colors.R
\docType{data}
\name{sano_colors}
\alias{sano_colors}
\title{Colors of Sciensano}
\format{
\subsection{\code{Sciensano}}{

A list with the different color palettes of Sciensano
\describe{
This data contains all hex values of the Sciensano color palette
}
}
}
\usage{
sano_colors
}
\description{
Colors of Sciensano
}
\keyword{datasets}
