% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_sano_plot.R
\name{save_sano_plot}
\alias{save_sano_plot}
\title{Title}
\usage{
save_sano_plot(plot_name, plot, height = 1000, width = 1500, source = NULL)
}
\arguments{
\item{plot_name}{the name of your exported plot and format as "plot.jpg"}

\item{plot}{the plot object}

\item{height}{the height in pixels}

\item{width}{the width in pixels}

\item{source}{the data source}
}
\value{
save your plot in a sano style
}
\description{
Title
}
\examples{
\dontrun{
save_sano_plot("plot.jpg",plot)
}
}
