% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_sano_plot.R
\name{style_sano_plot}
\alias{style_sano_plot}
\title{Apply Sano Styling to a ggplot Object}
\usage{
style_sano_plot(
  x,
  gridlines_horizontal = TRUE,
  gridlines_vertical = TRUE,
  font = "Arial"
)
}
\arguments{
\item{x}{A ggplot object to be styled.}

\item{gridlines_horizontal}{Logical. If \code{TRUE}, displays horizontal gridlines. Defaults to \code{TRUE}.}

\item{gridlines_vertical}{Logical. If \code{TRUE}, displays vertical gridlines. Defaults to \code{TRUE}.}

\item{font}{Character. The font family to use for text elements. Defaults to \code{"Arial"}.}
}
\value{
A ggplot object with the Sano styling applied.
}
\description{
This function customizes a ggplot object by applying the Sano style, which includes specific font settings, gridline options, and other theme adjustments.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()
styled_p <- style_sano_plot(p)
print(styled_p)
}
}
