% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_sano_table.R
\name{style_sano_table}
\alias{style_sano_table}
\title{style_sano_table() is using the flextable package to style your tables.
It can be used within any type of Rmarkdown.}
\usage{
style_sano_table(x, caption = NULL, autofit = F, scroll = T)
}
\arguments{
\item{x}{a table}

\item{caption}{a caption}

\item{autofit}{default to FALSE, set to TRUE for page fitting tables in Word Rmd.}

\item{scroll}{default to TRUE, adds a scroll for tables bigger than 20 rows (only in HTLM format).}
}
\value{
a Sciensano styled table
}
\description{
style_sano_table() is using the flextable package to style your tables.
It can be used within any type of Rmarkdown.
}
\examples{
style_sano_table(head(mtcars))

}
